+++ bepensa-models/DTO/ActualizarConctactosDTO.cs
@@ -7,7 +7,7 @@ namespace bepensa_models.DTO
     {
         [Display(Name = "IdUsuario")]
         [Required(ErrorMessage = "El campo {0} es obligatorio.")]
-        public long IdUsuario { get; set; }
+        public int IdUsuario { get; set; }
 
         [Display(Name = "Correo electrónico")]
         [Required(ErrorMessage = "El campo {0} es obligatorio.")]
